+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - ***BREAKING*** Updated supported `jj` version to 0.44.
+- Synced op heads are now made visible *after* the index has been built. This prevent user's `jj`
+  commands to block when indexing is in progress.
 - The Nix flake now uses `buildRustPackage` instead of [crane](https://crane.dev/). It no longer
   leaves large build artifacts in the user nix store after compiling from source.